Moved dragon_editor folder into the screens folder
